Switch to static version numbering
