+++ CveXplore/__init__.py
@@ -1,8 +1,1 @@
 from CveXplore.main import CveXplore
-
-try:
-    from version import _version
-
-    _version()
-except ModuleNotFoundError:
-    pass
