+++ CveXplore/main.py
@@ -51,13 +51,14 @@ from CveXplore.errors.validation import CveNumberValidationError
 from CveXplore.objects.cvexplore_object import CveXploreObject
 
 
-try:
-    from version import VERSION
-except ModuleNotFoundError:
+def _version():
     _PKG_DIR = os.path.dirname(__file__)
     version_file = os.path.join(_PKG_DIR, "VERSION")
     with open(version_file, "r") as fdsec:
         VERSION = fdsec.read()
+
+
+VERSION = __version__ = _version()
 
 
 class CveXplore(object):
