+++ setup.py
@@ -1,15 +1,22 @@
 import os
+import re
 from itertools import chain
 
 from setuptools import setup, find_packages
 
-try:
-    from version import VERSION
-except ModuleNotFoundError:
+
+# Get static version from CveXplore/VERSION
+def _version():
     _PKG_DIR = os.path.dirname(__file__)
-    version_file = os.path.join(_PKG_DIR, "CveXplore", "VERSION")
+    version_file = os.path.join(_PKG_DIR, "CveXplore/VERSION")
+
     with open(version_file, "r") as fdsec:
-        VERSION = fdsec.read()
+        tag = fdsec.read()
+    return tag
+
+
+VERSION = __version__ = _version()
+VERSION_MAIN = re.search(r"[0-9.]+", VERSION).group()
 
 # The directory containing this file
 HERE = os.path.abspath(os.path.dirname(__file__))
